requirements.txt
aiohttp==2.2.3
async-timeout==1.2.1
certifi==2017.4.17
chardet==3.0.4
click==6.7
future==0.16.0
idna==2.5
multidict==3.1.1
oauthlib==2.0.2
olefile==0.44
Pillow==4.2.1
pyglet==1.2.4
python-twitter==3.3
requests==2.18.1
requests-oauthlib==0.8.0
simplejson==3.11.1
six==1.10.0
uritemplate==3.0.0
urllib3==1.21.1
yarl==0.12.0
